Minor modifications to IcNumber and Birthday attributes to add javadocs and align with coding standards
